added Cpt, Bore and BaseParser classes
