+++ pygef/__init__.py
@@ -2,3 +2,5 @@ from pygef._version import __version__
 from pygef.gef import ParseGEF
 from pygef.plot_utils import plot_merged_cpt_bore
 from pygef.utils import depth_to_nap, join_gef, nap_to_depth
+from pygef.cpt import Cpt
+from pygef.bore import Bore
