+++ pygef/broxml.py
@@ -1,0 +1,3 @@
+class ParseBroXml:
+    def __init__(self, path=None, string=None):
+        pass
